requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
aio-pika==9.5.5
aioredis==2.0.1
pydantic>=1.8.0
python-dotenv>=0.19.0
locust==2.34.1
psycopg2-binary==2.9.10
sqlalchemy>=2.0.0
asyncpg>=0.27.0
alembic>=1.10.2
uvicorn==0.34.1
pydantic-settings==2.8.1
greenlet==3.1.1
typing-extensions>=4.6.0
